tests/wpt/tests/html/canvas/offscreen/line-styles/2d.line.join.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.line.join.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.line.join.valid</h1>
<p class="desc">Setting lineJoin to valid values works</p>


<script>
var t = async_test("Setting lineJoin to valid values works");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.lineJoin = 'bevel'
  _assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

  ctx.lineJoin = 'round';
  _assertSame(ctx.lineJoin, 'round', "ctx.lineJoin", "'round'");

  ctx.lineJoin = 'miter';
  _assertSame(ctx.lineJoin, 'miter', "ctx.lineJoin", "'miter'");
  t.done();

});
</script>
